SEE COMMIT DESCRIPTION, added stuff to help testing
IMPORTANT: in the particle class, there was a warning that said when
implementing Comparable, it needs to be done like this:
"Comparable<Particle>", not just "Comparable"
Then after doing that, the compareTo method needed to be changed to:
"public int compareTo(Particle Par)",
not "public int compareTo(Object o)".
Do note, it was just a warning, not an error, so when we wrote the
compare stuff on Saturday we ignored it. But...
This might fix that problem where the program "Violated its general
contract" when comparing, which if it did that would be very nice, so we
should go back and test this to see if it works.
Also added some stuff to help test with the current code